Update a URL to the Lithuanian Braille standard
The old URL pointed to the 8-dot specification document, but that system is barely used. The new URL points to an actual standard which specifies how Braille is to be used in numerous contexts (literary, math, music, chess). Sadly, the standard is only available for purchase and may not be redistributed legally.